댓글 어카운트 지정 및 권한 부여

+++ backend/community/views.py
@@ -72,6 +72,7 @@ def post(request, pk):
     if request.method == "POST":
         newComment = post.comment_set.create(
             body=data["body"], pubDate=timezone.now())
+        newComment.author = request.account
         newComment = model_to_dict(newComment)
         result = JsonResponse(newComment)
         return result
@@ -87,6 +88,8 @@ def comment(request, pk, comment_id):
 
     # 댓글 수정
     if request.method == "PATCH":
+        if request.account != comment.author:
+            return JsonResponse("사용자의 글이 아닙니다", safe=False)
         comment.body = data["body"]
         comment.save()
         result = JsonResponse(model_to_dict(comment))
@@ -94,5 +97,7 @@ def comment(request, pk, comment_id):
 
     # 댓글 삭제
     if request.method == "DELETE":
+        if request.account != comment.author:
+            return JsonResponse("사용자의 글이 아닙니다", safe=False)
         comment.delete()
         return JsonResponse(model_to_dict(comment))
